+++ snippet.DocLister.php
@@ -4,8 +4,8 @@
  *
  * @license GNU General Public License (GPL), http://www.gnu.org/copyleft/gpl.html
  * @author Agel_Nash <Agel_Nash@xaker.ru>
- * @date 03.05.2013
- * @version 1.0.12
+ * @date 20.08.2013
+ * @version 1.0.20
  */
 if (!defined('MODX_BASE_PATH')) {
     die('HACK???');